Fix IntegrityError on double answer submission
When an employer clicks twice on an answer form submit button, there is
the possibility of a race condition where we attempt to create two
different Answer objects for the same job application. This causes a 500
error that we fix by catching the error and displaying a "already
answered" message.