Fix missing exceptions from Sentry
In sentry, exceptions with stacktraces bigger than 200kb are not
processed (413 error). This occurs frequently, and when we need it the
most: when email attachments are too large. To address this, we log an
error in addition to exceptions where it is required. Errors do not
contain a stack trace, but at least they warn the developers that an
error occur.