Process email events
Our emailing provider (Mailjet, Postal, or other) sends us events
related to emails. Among these, it is crucial that we process
spam/blocked/bounce events.

To do so, we first switch from the SMTP API to the HTTP API. This allows
us to collect message IDs that we use to track the life the emails.

Then, after each event received, we trigger an asynchronous event. For
spam/blocked events, we add the guilty address to a cache with an
expiration date of 1 month. That means that it should not be possible to
apply to that company, or to send applications from this address, for a
duration of 1 month.

Note that here we do not warn recipients that their previous
applications were not received, this is work for another time...